+++ CHANGELOG.md
@@ -2,4 +2,9 @@
 # 1.0.0 (2017-12-22)
 
 
+### Bug Fixes
+
+* fix return function ([5eb0f92](https://github.com/wangliangyu/aqi-vue/commit/5eb0f92))
+
+